Move promotion-specific code from product.rb to gem
We need to purge references to Solidus' legacy promotion system from the
core.
